badgereader_addon/requirements.txt
PyYAML
aiohttp
gspread
